+++ tests/test_analysis.py
@@ -14,13 +14,6 @@ from pheval.analyse.analysis import (
     parse_pheval_gene_result,
     parse_pheval_variant_result,
 )
-from pheval.analyse.generate_plots import (
-    PlotGenerator,
-    TrackGenePrioritisation,
-    TrackPrioritisation,
-    TrackVariantPrioritisation,
-)
-from pheval.analyse.generate_summary_outputs import RankComparisonGenerator, merge_results
 from pheval.analyse.rank_stats import RankStats
 from pheval.post_processing.post_processing import (
     PhEvalGeneResult,
@@ -249,61 +242,6 @@ class TestPrioritisationRankRecorder(unittest.TestCase):
                 },
             ),
         )
-
-
-class TestRankStats(unittest.TestCase):
-    def setUp(self) -> None:
-        self.rank_stats = RankStats()
-
-    def test_add_rank(self):
-        self.rank_stats.add_rank(1)
-        self.rank_stats.add_rank(3)
-        self.rank_stats.add_rank(5)
-        self.rank_stats.add_rank(7)
-        self.rank_stats.add_rank(10)
-        self.assertEqual(
-            self.rank_stats,
-            RankStats(
-                top=1,
-                top3=2,
-                top5=3,
-                top10=5,
-                found=5,
-                total=0,
-                reciprocal_ranks=[1.0, 0.3333333333333333, 0.2, 0.14285714285714285, 0.1],
-            ),
-        )
-
-    def test_percentage_rank(self):
-        self.rank_stats.found = 10
-        self.assertTrue(self.rank_stats.percentage_rank(3) == 30)
-
-    def test_percentage_top(self):
-        self.rank_stats.top, self.rank_stats.found = 10, 20
-        self.assertEqual(self.rank_stats.percentage_top(), 50)
-
-    def test_percentage_top3(self):
-        self.rank_stats.top3, self.rank_stats.found = 30, 50
-        self.assertEqual(self.rank_stats.percentage_top3(), 60)
-
-    def test_percentage_top5(self):
-        self.rank_stats.top5, self.rank_stats.found = 70, 160
-        self.assertEqual(self.rank_stats.percentage_top5(), 43.75)
-
-    def test_percentage_top10(self):
-        self.rank_stats.top10, self.rank_stats.found = 100, 160
-        self.assertEqual(self.rank_stats.percentage_top10(), 62.5)
-
-    def test_percentage_found(self):
-        self.rank_stats.found, self.rank_stats.total = 100, 125
-        self.assertEqual(self.rank_stats.percentage_found(), 80)
-
-    def test_percentage_difference(self):
-        self.assertEqual(self.rank_stats.percentage_difference(54, 23), 31)
-
-    def test_mean_reciprocal_rank(self):
-        self.rank_stats.reciprocal_ranks = [0.2, 0.4, 0.5, 0.6, 0.8]
-        self.assertEqual(self.rank_stats.mean_reciprocal_rank(), 0.5)
 
 
 class TestAssessGenePrioritisation(unittest.TestCase):
@@ -925,125 +863,6 @@ class TestAssessVariantPrioritisation(unittest.TestCase):
         )
 
 
-class TestMergeResults(unittest.TestCase):
-    def setUp(self) -> None:
-        self.result_1 = {
-            1: {
-                "Phenopacket": "phenopacket1.json",
-                "Gene": "GCDH",
-                "/path/to/results_directory1": 1,
-            }
-        }
-        self.result_2 = {
-            1: {
-                "Phenopacket": "phenopacket1.json",
-                "Gene": "GCDH",
-                "/path/to/results_directory2": 5,
-            }
-        }
-
-    def test_merge_results(self):
-        self.assertEqual(
-            merge_results(self.result_1, self.result_2),
-            {
-                1: {
-                    "Phenopacket": "phenopacket1.json",
-                    "Gene": "GCDH",
-                    "/path/to/results_directory1": 1,
-                    "/path/to/results_directory2": 5,
-                }
-            },
-        )
-
-
-class TestRankComparisonGenerator(unittest.TestCase):
-    def setUp(self) -> None:
-        self.gene_rank_comparisons = RankComparisonGenerator(
-            defaultdict(
-                dict,
-                {
-                    1: {
-                        "Phenopacket": "phenopacket1.json",
-                        "Gene": "GCDH",
-                        "/path/to/results_directory1": 1,
-                        "/path/to/results_directory2": 5,
-                    }
-                },
-            )
-        )
-        self.variant_rank_comparisons = RankComparisonGenerator(
-            defaultdict(
-                dict,
-                {
-                    1: {
-                        "Phenopacket": "phenopacket1.json",
-                        "Variant": "3_12563453454_C_T",
-                        "/path/to/results_directory1": 9,
-                        "/path/to/results_directory2": 3,
-                    }
-                },
-            )
-        )
-
-    def test_generate_gene_dataframe(self):
-        result = pd.DataFrame(
-            [
-                {
-                    "Phenopacket": "phenopacket1.json",
-                    "Gene": "GCDH",
-                    "/path/to/results_directory1": 1,
-                    "/path/to/results_directory2": 5,
-                }
-            ]
-        )
-        result.index += 1
-        self.assertTrue(result.equals(self.gene_rank_comparisons._generate_dataframe()))
-
-    def test_generate_variant_dataframe(self):
-        result = pd.DataFrame(
-            [
-                {
-                    "Phenopacket": "phenopacket1.json",
-                    "Variant": "3_12563453454_C_T",
-                    "/path/to/results_directory1": 9,
-                    "/path/to/results_directory2": 3,
-                }
-            ]
-        )
-        result.index += 1
-        self.assertTrue(result.equals(self.variant_rank_comparisons._generate_dataframe()))
-
-    def test_calculate_gene_rank_difference(self):
-        result = pd.DataFrame(
-            [
-                {
-                    "Phenopacket": "phenopacket1.json",
-                    "Gene": "GCDH",
-                    "/path/to/results_directory1": 1,
-                    "/path/to/results_directory2": 5,
-                    "rank_decrease": 4,
-                }
-            ]
-        )
-        result.index += 1
-        self.assertTrue(result.equals(self.gene_rank_comparisons._calculate_rank_difference()))
-
-    def test_calculate_variant_rank_difference(self):
-        result = pd.DataFrame(
-            [
-                {
-                    "Phenopacket": "phenopacket1.json",
-                    "Variant": "3_12563453454_C_T",
-                    "/path/to/results_directory1": 9,
-                    "/path/to/results_directory2": 3,
-                    "rank_decrease": -6,
-                }
-            ]
-        )
-        result.index += 1
-        self.assertTrue(result.equals(self.variant_rank_comparisons._calculate_rank_difference()))
-
-
 class TestParsePhEvalGeneResult(unittest.TestCase):
     def test_parse_pheval_gene_result(self):
         self.assertEqual(
@@ -1179,129 +998,5 @@ class TestParsePhEvalVariantResult(unittest.TestCase):
                     ),
                     rank=1,
                 ),
-            ],
-        )
-
-
-class TestPlotGenerator(unittest.TestCase):
-    def setUp(self) -> None:
-        self.gene_plot_generator = PlotGenerator(gene_analysis=True)
-        self.variant_plot_generator = PlotGenerator(gene_analysis=False)
-        self.track_prioritisation = TrackPrioritisation(
-            gene_prioritisation=TrackGenePrioritisation(
-                results_dir=Path("/path/to/tool/corpus_results"),
-                ranks={},
-                rank_stats=RankStats(
-                    top=1,
-                    top3=2,
-                    top5=3,
-                    top10=9,
-                    found=20,
-                    total=30,
-                    reciprocal_ranks=[1, 1 / 3, 1 / 5, 1 / 10, 1 / 50],
-                ),
-            ),
-            variant_prioritisation=TrackVariantPrioritisation(
-                results_dir=Path("/path/to/tool/corpus_results"),
-                ranks={},
-                rank_stats=RankStats(
-                    top=1,
-                    top3=2,
-                    top5=3,
-                    top10=4,
-                    found=5,
-                    total=2,
-                    reciprocal_ranks=[1, 1 / 3, 1 / 5, 1 / 10, 1 / 12],
-                ),
-            ),
-        )
-
-    def test__retrieve_prioritisation_data_gene(self):
-        self.assertEqual(
-            self.gene_plot_generator._retrieve_prioritisation_data(self.track_prioritisation),
-            TrackGenePrioritisation(
-                results_dir=Path("/path/to/tool/corpus_results"),
-                ranks={},
-                rank_stats=RankStats(
-                    top=1,
-                    top3=2,
-                    top5=3,
-                    top10=9,
-                    found=20,
-                    total=30,
-                    reciprocal_ranks=[1, 0.3333333333333333, 0.2, 0.1, 0.02],
-                ),
-            ),
-        )
-
-    def test__retrieve_prioritisation_data_variant(self):
-        self.assertEqual(
-            self.variant_plot_generator._retrieve_prioritisation_data(self.track_prioritisation),
-            TrackVariantPrioritisation(
-                results_dir=PosixPath("/path/to/tool/corpus_results"),
-                ranks={},
-                rank_stats=RankStats(
-                    top=1,
-                    top3=2,
-                    top5=3,
-                    top10=4,
-                    found=5,
-                    total=2,
-                    reciprocal_ranks=[1, 0.3333333333333333, 0.2, 0.1, 0.08333333333333333],
-                ),
-            ),
-        )
-
-    def test__generate_stacked_bar_plot_data(self):
-        self.gene_plot_generator._generate_stacked_bar_plot_data(self.track_prioritisation)
-        self.assertEqual(
-            self.gene_plot_generator.stats,
-            [
-                {
-                    "2-3": 5.0,
-                    "4-5": 5.0,
-                    "6-10": 30.0,
-                    ">10": 21.66666666666667,
-                    "FO/NP": 33.33333333333333,
-                    "Run": "tool_corpus_results",
-                    "Top": 5.0,
-                }
-            ],
-        )
-
-    def test__generate_stats_mrr_bar_plot_data(self):
-        self.gene_plot_generator._generate_stats_mrr_bar_plot_data(self.track_prioritisation)
-        self.assertEqual(
-            self.gene_plot_generator.mrr,
-            [{"Rank": "MRR", "Percentage": 0.33066666666666666, "Run": "tool_corpus_results"}],
-        )
-
-    def test__generate_cumulative_bar_plot_data(self):
-        self.gene_plot_generator._generate_cumulative_bar_plot_data(self.track_prioritisation)
-        self.assertEqual(
-            self.gene_plot_generator.stats,
-            [
-                {"Rank": "Top", "Percentage": 0.05, "Run": "tool_corpus_results"},
-                {"Rank": "Top3", "Percentage": 0.1, "Run": "tool_corpus_results"},
-                {"Rank": "Top5", "Percentage": 0.15, "Run": "tool_corpus_results"},
-                {"Rank": "Top10", "Percentage": 0.45, "Run": "tool_corpus_results"},
-                {"Rank": "Found", "Percentage": 0.6666666666666667, "Run": "tool_corpus_results"},
-                {"Rank": "FO/NP", "Percentage": 0.33333333333333326, "Run": "tool_corpus_results"},
-                {"Rank": "MRR", "Percentage": 0.33066666666666666, "Run": "tool_corpus_results"},
-            ],
-        )
-
-    def test__generate_non_cumulative_bar_plot_data(self):
-        self.gene_plot_generator._generate_non_cumulative_bar_plot_data(self.track_prioritisation)
-        self.assertEqual(
-            self.gene_plot_generator.stats,
-            [
-                {"Rank": "Top", "Percentage": 0.05, "Run": "tool_corpus_results"},
-                {"Rank": "2-3", "Percentage": 0.05, "Run": "tool_corpus_results"},
-                {"Rank": "4-5", "Percentage": 0.05, "Run": "tool_corpus_results"},
-                {"Rank": "6-10", "Percentage": 0.3, "Run": "tool_corpus_results"},
-                {"Rank": ">10", "Percentage": 0.2166666666666667, "Run": "tool_corpus_results"},
-                {"Rank": "FO/NP", "Percentage": 0.33333333333333326, "Run": "tool_corpus_results"},
-                {"Rank": "MRR", "Percentage": 0.33066666666666666, "Run": "tool_corpus_results"},
             ],
         )
